tests/requirements.txt
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
locust>=2.20.0
websocket-client>=1.7.0
